# docs/swagger.yaml
openapi: 3.0.1
info:
  title: My Node.js API
  description: API documentation for my Node.js app
  version: 1.0.0
servers:
  - url: http://localhost:{port}
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer

paths:
  /auth/register:
      post:
        summary: User registration endpoint. For role only "admin", "manager" or "developer" are allowed.
        requestBody:
          required: true
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    type: string
                    example: john_doe
                  password:
                    type: string
                    example: mypassword123
                  role:
                    type: string
                    enum: 
                      - admin
                      - manager
                      - developer
        responses:
          '201':
            description: User created successfully
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: User created successfully
          '500':
            description: Error creating user
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    message:
                      type: string
                      example: Error creating user
                    error:
                      type: string
                      example: Error details
  /auth/login:
    post:
      summary: User login endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  example: john_doe
                password:
                  type: string
                  example: mypassword123
      responses:
        '200':
          description: Successful login
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Invalid username or password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Invalid username or password.
  /projects:
    post:
      summary: Project creation endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: My Project
                description:
                  type: string
                  example: This is a project description.
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Project created successfully
                  project:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      name:
                        type: string
                        example: My Project
                      description:
                        type: string
                        example: This is a project description.
        '500':
          description: Error creating project
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error creating project
                  error:
                    type: string
                    example: Error details
    get:
      summary: List all projects
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Projects found successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Proyectos encontrados
                  list:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        name:
                          type: string
                          example: Project Name
                        description:
                          type: string
                          example: Project Description
        '500':
          description: Error finding projects
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error finding projects
                  error:
                    type: string
                    example: Error details
  /tasks:
    post:
      summary: Create a new task
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: Task Title
                description:
                  type: string
                  example: Task Description
                project_id:
                  type: integer
                  example: 1
                assigned_to:
                  type: string
                  example: john_doe
      responses:
        '201':
          description: Task created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Task created successfully
                  task:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      title:
                        type: string
                        example: Task Title
                      description:
                        type: string
                        example: Task Description
                      project_id:
                        type: integer
                        example: 1
                      assigned_to:
                        type: string
                        example: john_doe
        '500':
          description: Error creating task
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error creating task
                  error:
                    type: string
                    example: Error details
    get:
      summary: List all tasks
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tasks found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tareas encontradas
                  list:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          example: 1
                        title:
                          type: string
                          example: Task Title
                        description:
                          type: string
                          example: Task Description
                        project_id:
                          type: integer
                          example: 1
                        assigned_to:
                          type: string
                          example: john_doe
                        createdAt:
                          type: string
                          format: date-time
                          example: 2023-06-24T15:00:00Z
                        updatedAt:
                          type: string
                          format: date-time
                          example: 2023-06-24T15:00:00Z
        '500':
          description: Error finding tasks
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Error finding tasks
                  error:
                    type: string
                    example: Error details